requirements.txt
beautifulsoup4~=4.12.2
requests~=2.31.0
cloudscraper~=1.2.71
python-dotenv~=1.0.0
